Her Anders- Det virker lidt. FIX: Læs det ind som int => Cast det til char

+++ Enhedsregister V0.0.0.0.0.0.0.0.0.1/F16PRJ2/EnhedsRegister/EnhedsRegister/unit.cpp
@@ -3,8 +3,8 @@
 std::ostream& operator<<(std::ostream& os, const unit& obj)
 {
 	return os
-		<< obj.getUnitID()	<< " "
-		<< obj.getRoomID()	<< " "
+		<< static_cast<int>(obj.getUnitID() )	<< " "
+		<< static_cast<int>(obj.getRoomID() )	<< " "
 		<< obj.getStatus()	<< " "
 		<< obj.getTime()	<< " ";
 }
